add some StoreTest hooks into a higher level abstract class
